pyproject.toml
[project]
name = "dora-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Oleh Borysevych",email = "borysevych.oleh87@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dynamic = [ "dependencies" ]

[tool.poetry]
package-mode = false

[tool.poetry.requires-plugins]
poetry-plugin-shell = ">=1.0.1"

[tool.poetry.dependencies]
python-dotenv = "1.0.1"
langchain = "^0.3.25"
langchain-openai = "^0.3.16"
pydantic = "^2.11.4"
rich = "^14.0.0"
neo4j = "^5.28.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 140

[tool.isort]
py_version = 311
line_length = 140

